[Refactor]: slice 폴더 생성
slice만 모아 놓는 폴더를 만들고 파일을 이동시킵니다.
